requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pymongo==4.10.1
anthropic==0.72.0
python-dotenv==1.0.1
pydantic==2.10.3
google-genai>=0.7.0
